==> real_estate/v2/02_df_to_supervised.ipynb <==
import pandas as pd

df = pd.read_csv("~/real_estate/data/unsupervised_dfs/df_20181210.csv")

features = [
 'median_sqft_value',
 'percent_decreasing',
 'percent_increasing',
 'zhvi_singlefam',
 'zhvi_1bed',
 'zhvi_2bed',
 'zhvi_3bed',
 'zhvi_condo',
 '90_day_treasury_bill_rate',
 'mean_hourly_earnings_adj',
 'mean_fixed_mortgage_rate',
 'total_nonfarm_payroll',
 'zhvi_sqft_value']

target = "zhvi_sqft_value"

# Store data from previous, future months and years
feature_mo = {}
feature_yr = {}
zhvi_sqft_value_12_months = []

for i in features:
    feature_mo[i+"_mo"] = []
    feature_yr[i+"_yr"] = []

for i in df.index:
    
    current_row = df.iloc[i]
    region_name = current_row["RegionName"]
    month = int(current_row["month"])
    year = int(current_row["year"])
    next_year = year + 1
    
    # Get column values of previous time preiods
    last_year = year - 1
    if month == 1:
        last_month_month = 12
        last_month_year = year - 1
    else:
        last_month_month = month - 1
        last_month_year = year
    
    last_month_row = df.loc[(df["RegionName"] == region_name) & 
                            (df["month"] == last_month_month) &
                            (df["year"] == last_month_year)]
    
    last_year_row = df.loc[(df["RegionName"] == region_name) &
                           (df["month"] == month) &
                           (df["year"] == last_year)]
    
    next_year_row = df.loc[(df["RegionName"] == region_name) &
                           (df["month"] == month) &
                           (df["year"] == next_year)]
    
    for i in features:
        
        feature_mo[i+"_mo"].append(last_month_row[i].values)
        feature_yr[i+"_yr"].append(last_year_row[i].values)
        
    next_year_target = next_year_row["zhvi_sqft_value"]
    
    if len(next_year_target) == 0:
        next_year_target = None
    else:
        next_year_target = next_year_target.values[0]
    
    zhvi_sqft_value_12_months.append(next_year_target)
    
# Add data to dataframe
for i in feature_mo.keys():
    new_list = [j[0] if len(j) > 0 else None for j in feature_mo[i]]
    feature_mo[i] = new_list

for i in feature_yr.keys():
    new_list = [j[0] if len(j) > 0 else None for j in feature_yr[i]]
    feature_yr[i] = new_list
    
for i in feature_mo.keys():
    df[i] = feature_mo[i]
    
for i in feature_yr.keys():
    df[i] = feature_yr[i]
    
df["zhvi_sqft_value_12_months"] = zhvi_sqft_value_12_months

# Filter null values
df = df.dropna()

# Export
df.to_csv("~/real_estate/data/supervised_dfs/df_20181210.csv", index=False)

df.head()
